fix(sitemap,seo): prevent sitemap 504s and give program pages a self-canonical
The per-kind sitemap routes generate on demand by paging the indexer; cold,
the projects sitemap already takes ~10s to assemble. None of the sitemap
routes set maxDuration, so they ran against the platform default (~10s) and
intermittently 504 — exactly what Search Console reports as 'Couldn't fetch'.
Add maxDuration=60 to every sitemap route (and force-dynamic to the two [kind]
routes that lacked it) so a cold build has headroom; the warmer keeps the Data
Cache hot so that duration is only ever spent off the crawler path.

Funding-program detail pages had no generateMetadata, so they inherited the
community layout's canonical (/community/<id>) — Google folded every program
into the community root and none of the funding-programs sitemap URLs indexed
as themselves. Split the page into a server component that emits a
self-referential canonical plus real title/description, and a client component
holding the existing UI.
