+++ app/sitemap-index.xml/route.ts
@@ -4,6 +4,8 @@ import { sitemapIndexResponse } from "@/utilities/sitemap";
 // Never prerender at build time — render on demand only. The indexer fetch is
 // still cached via the Data Cache (see fetchSitemapCounts).
 export const dynamic = "force-dynamic";
+// Headroom for the cold counts fetch so the index itself never 504s.
+export const maxDuration = 60;
 
 // Legacy index URL. Google's stored state for it is stuck on a degraded May
 // 2026 parse, so robots.txt now advertises /sitemap_index.xml instead (see
